Split ServiceFileTransformerTest
